Update PPP, Developer Guide and restructure Document and inheriting classes
