+++ src/main/java/seedu/address/logic/commands/ReceiptCommand.java
@@ -43,7 +43,7 @@ public class ReceiptCommand extends QueueCommand {
     @Override
     public CommandResult execute(Model model, PatientQueue patientQueue, CurrentPatient currentPatient,
                                  ServedPatientList servedPatientList, CommandHistory history) throws CommandException {
-        requireNonNull(model);
+        requireNonNull(servedPatientList);
 
         if (index.getZeroBased() >= servedPatientList.getServedPatientListLength()) {
             throw new CommandException(Messages.MESSAGE_INVALID_PERSON_DISPLAYED_INDEX);
