+++ src/main/java/seedu/address/model/document/ReferralLetter.java
@@ -9,6 +9,9 @@ import seedu.address.model.person.ServedPatient;
 public class ReferralLetter extends Document {
     public static final String FILE_TYPE = "Referral Letter";
 
+    private String referralContent;
+    private String noteContent;
+
     /**
      * Creates a ReferralLetter object for the specified servedPatient.
      * @param servedPatient the patient who has already consulted the doctor.
@@ -17,7 +20,29 @@ public class ReferralLetter extends Document {
         setFileType(FILE_TYPE);
         setName(servedPatient.getName());
         setIcNumber(servedPatient.getIcNumber());
-        setReferralContent(servedPatient.getReferralContent());
-        setNoteContent(servedPatient.getNoteContent());
+        referralContent = servedPatient.getReferralContent();
+        noteContent = servedPatient.getNoteContent();
+    }
+
+    public String formatInformation() {
+        String referralContent = getReferralContent();
+        String noteContent = getNoteContent();
+
+        StringBuilder stringBuilder = new StringBuilder();
+        stringBuilder.append("This is to certify that the above-named patient has been referred to: ")
+                .append("<b>" + referralContent.toUpperCase() + "</b>" + "<br><br>")
+                .append("Dear Specialist, please assist the above-named patient in the following matter:<br>")
+                .append(noteContent + "<br><br>")
+                .append("Kindly do accept him under your care. Thank you very much.<br><br>")
+                .append("<b>Issuing Doctor:</b> Dr Chester Sng" + "<br>");
+        return stringBuilder.toString();
+    }
+
+    public String getReferralContent() {
+        return referralContent;
+    }
+
+    public String getNoteContent() {
+        return noteContent;
     }
 }
