+++ src/main/resources/view/Documents/Receipt_For_Bryan_S9637283X.html
@@ -14,7 +14,7 @@
     </script>
     <h1>
         <img src = "../../images/CLInicLogo.PNG" alt = "CLInic's logo" align = "right" height="70px" width="350px">
-        <div class = "header" id="headerDiv">Receipt<br>Issued On: 01/11/2018 23:08:52</div>
+        <div class = "header" id="headerDiv">Receipt<br>Issued On: 05/11/2018 23:59:36</div>
     </h1>
 </head>
 <div class="main">
@@ -29,7 +29,7 @@
         </tr>
     </table>
     <br>
-    <span id="contentDiv"><table ID = "contentTable" width = 100%><col width = "700"><tr ID = "receiptHeader"><div class="contentHeader"><th>Prescription</th><th>Quantity</th><th>Unit Price</th><th>Total Price</th></div></tr><tr ID = "receiptEnd"><td>Grand Total:</td><td>-</td><td>-</td><td>0.00</td></tr></table></span>
+    <span id="contentDiv"><table ID = "contentTable" width = 100%><col width = "700"><tr ID = "receiptHeader"><div class="contentHeader"><th>Prescription</th><th>Quantity</th><th>Unit Price</th><th>Total Price</th></div></tr><tr><td>CONSULTATION</td><td>1</td><td>30.00</td><td>30.00</td></tr><tr><td>lol</td><td>60</td><td>2.00</td><td>120.00</td></tr><tr ID = "receiptEnd"><td>Grand Total:</td><td>-</td><td>-</td><td>150.00</td></tr></table></span>
     <br>
     <span style=" display: block; text-align: center">Have a nice day!</span>
 </div>
