+++ src/main/resources/view/Documents/Referral_Letter_For_Bryan_S9637283X.html
@@ -14,7 +14,7 @@
     </script>
     <h1>
         <img src = "../../images/CLInicLogo.PNG" alt = "CLInic's logo" align = "right" height="70px" width="350px">
-        <div class = "header" id="headerDiv">Referral Letter<br>Issued On: 01/11/2018 23:08:56</div>
+        <div class = "header" id="headerDiv">Referral Letter<br>Issued On: 05/11/2018 23:59:46</div>
     </h1>
 </head>
 <div class="main">
@@ -29,7 +29,7 @@
         </tr>
     </table>
     <br>
-    <span id="contentDiv">Dear Specialist, please assist the above-named patient in the following matter:<br>yolosweg<br><br>Kindly do accept him under your care. Thank you very much.<br><br><b>Issuing Doctor:<b> DR CHESTER SNG<br><br></span>
+    <span id="contentDiv">This is to certify that the above-named patient has been referred to: <b>HELP PLS</b><br><br>Dear Specialist, please assist the above-named patient in the following matter:<br>dying<br><br>Kindly do accept him under your care. Thank you very much.<br><br><b>Issuing Doctor:</b> Dr Chester Sng<br></span>
     <br>
     <span style=" display: block; text-align: center">Have a nice day!</span>
 </div>
@@ -76,6 +76,5 @@
     }
     #contentTable tr:nth-child(even) {
         background-color: #f2f2f2;
-        border-style: groove;
     }
 </style>
