--- src/test/java/seedu/address/model/document/DocumentTest.java
@@ -1,4 +1,0 @@
-package seedu.address.model.document;
-
-public class DocumentTest {
-}
